dh w refactor sesuai yg lo bilg
