+++ src/components/test/BackButton.tsx
@@ -1,22 +1,37 @@
-import { Button } from "@chakra-ui/react";
+import { Box, Text } from "@chakra-ui/react";
 
 interface BackButtonProps {
   onClick: () => void;
-  disabled?: boolean;
 }
 
-export const BackButton = ({ onClick, disabled = false }: BackButtonProps) => {
+export const BackButton = ({ onClick }: BackButtonProps) => {
   return (
-    <Button
-      variant="ghost"
-      colorPalette="gray"
+    <Box
+      display="flex"
+      alignItems="center"
+      gap={2}
+      cursor="pointer"
       onClick={onClick}
-      disabled={disabled}
-      size="md"
-      fontWeight="medium"
+      transition="all 0.2s"
+      color="gray.700"
+      _hover={{
+        color: "blue.500",
+      }}
+      mb={6}
+      width="fit-content"
     >
-      ← Sebelumnya
-    </Button>
+      <Box
+        display="flex"
+        alignItems="center"
+        justifyContent="center"
+        fontSize="xl"
+      >
+        ←
+      </Box>
+      <Text fontSize="md" fontWeight="medium">
+        Kembali
+      </Text>
+    </Box>
   );
 };
