+++ src/components/test/index.ts
@@ -2,4 +2,6 @@ export { ProgressBar } from "./ProgressBar";
 export { QuestionCard } from "./QuestionCard";
 export { ChoiceButton } from "./ChoiceButton";
 export { BackButton } from "./BackButton";
+export { FinishButton } from "./FinishButton";
 export { LoadingOverlay } from "./LoadingOverlay";
+export { default as TestContainer } from "./TestContainer";
